Move FEATURE_NAMES to encoding.schema (from definitions)
