+++ kissim/encoding/definitions.py
@@ -6,26 +6,6 @@ from pathlib import Path
 
 import pandas as pd
 
-FEATURE_NAMES = {
-    "physicochemical": [
-        "size",
-        "hbd",
-        "hba",
-        "charge",
-        "aromatic",
-        "aliphatic",
-        "sco",
-        "exposure",
-    ],
-    "distances": [
-        "distance_to_centroid",
-        "distance_to_hinge_region",
-        "distance_to_dfg_region",
-        "distance_to_front_pocket",
-    ],
-    "moments": ["moment1", "moment2", "moment3"],
-}
-
 SITEALIGN_FEATURES = pd.read_csv(
     Path(__name__).parent / "kissim" / "data" / "sitealign_features.csv", index_col=0
 )
