+++ kissim/encoding/schema.py
@@ -1,0 +1,23 @@
+"""
+kissim.encoding.schema TODO
+"""
+
+FEATURE_NAMES = {
+    "physicochemical": [
+        "size",
+        "hbd",
+        "hba",
+        "charge",
+        "aromatic",
+        "aliphatic",
+        "sco",
+        "exposure",
+    ],
+    "distances": [
+        "distance_to_centroid",
+        "distance_to_hinge_region",
+        "distance_to_dfg_region",
+        "distance_to_front_pocket",
+    ],
+    "moments": ["moment1", "moment2", "moment3"],
+}
