+++ kissim/similarity/api.py
@@ -15,7 +15,7 @@ import numpy as np
 import pandas as pd
 from scipy.spatial import distance
 
-from kissim.encoding.definitions import FEATURE_NAMES
+from kissim.encoding.schema import FEATURE_NAMES
 
 logger = logging.getLogger(__name__)
